pyproject.toml
[project]
name = "waycore-bank-scraper"
version = "0.1.0"
description = "Durable bank data extraction via browser automation and LLM"
requires-python = ">=3.12"
dependencies = [
    "restate-sdk>=0.9.0",
    "playwright>=1.49.0",
    "playwright-stealth>=1.0.6",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "cryptography>=44.0.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "hypercorn>=0.17.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
s3 = ["aioboto3>=13.0.0"]
all = ["anthropic>=0.40.0", "openai>=1.0.0", "aioboto3>=13.0.0"]

[project.scripts]
waycore = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ASYNC", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
